feat: auth refactor login and landingpage
