+++ client/app/components/customsComponents/layout/Header.tsx
@@ -1,7 +1,6 @@
 import * as React from "react";
 import Link from "next/link";
 import LogoLetter from "../design/LogoLetter";
-import { NavMenu } from "@/app/enums/NavMenuEnum";
 import Avatar from "@/app/userProfile/components/Avatar";
 
 interface HeaderProps {
@@ -19,13 +18,9 @@ export function Header({ title }: HeaderProps) {
                     <LogoLetter />
                 </Link>
                 <h1 className="absolute left-1/2 font-bold transform -translate-x-1/2 text-xl sm:text-2xl md:text-1xl">{title}</h1>
-                {title !== NavMenu.HOME ? (
-                    <Link href='/userProfile'>
-                        <Avatar src={profileImgSrc} alt="Avatar" />
-                    </Link>
-                ) : (
-                    <div className="w-6"></div> // Espace réservé pour le bouton "Settings"
-                )}
+                <Link href='/userProfile'>
+                    <Avatar src={profileImgSrc} alt="Avatar" />
+                </Link>
             </div>
         </header>
     );
